resuelto problema de la vista dinamica de Team, test correctos

+++ app/Http/Controllers/TransparencyController.php
@@ -14,7 +14,7 @@ class TransparencyController extends Controller
     }
     public function index()
     {
-
+        $this->authorize('view-any', Transparency::class);
         $transparencies = Transparency::paginate(8);
         if (auth()->user()->role_id != "Admin") {
             return view('user.notauthorized');
@@ -24,14 +24,15 @@ class TransparencyController extends Controller
 
     public function create()
     {
+        $this->authorize('create', Transparency::class);
         $transparencies = Transparency::all();
         return view('transparency.create', compact('transparencies'));
 
     }
 
-    public function store(Request $request)
+    public function store(Request $request, Transparency $transparency)
     {
-
+        $this->authorize('create', $transparency);
         $data = $request->all();
         $validatedData = $request->validate([
             'economic_document'=> 'max:22560',
@@ -45,16 +46,19 @@ class TransparencyController extends Controller
 
     public function show(Transparency $transparency)
     {
+        $this->authorize('view', Transparency::class);
         return view('transparency.show',compact('transparency'));
     }
 
     public function edit(Transparency $transparency)
     {
+        $this->authorize('update', Transparency::class);
         return view('transparency.edit',compact('transparency'));
     }
 
     public function update(Request $request, Transparency $transparency)
     {
+        $this->authorize('update', $transparency);
         $data = $request->all();
         $validatedData = $request->validate([
             'economic_document'=> 'max:22560',
@@ -68,6 +72,7 @@ class TransparencyController extends Controller
 
     public function destroy(Transparency $transparency)
     {
+        $this->authorize('destroy', $transparency);
         $transparency->delete();
         Storage::delete(['entity_document', 'economic_document']);
         return redirect()->route('transparency.index');
